Assert: Hedge
