+++ FL_TESTS
@@ -2,6 +2,7 @@ test/AssertActivation.h
 test/AssertConstructor.h
 test/AssertDefuzzifier.h
 test/AssertFactory.h
+test/AssertHedge.h
 test/Headers.h
 test/Listener.h
 test/Mock.h
